Chore: simply formatting with prettier
The configuration to ignore files by Prettier was purely for the issue
of the pre-commit hook that failed, if staged files should be formatted
that are not supported by Prettier. Because in these cases Prettier gets
called with the list of staged files. For the script in `package.json`
this does not happen because files are selected by a pattern.

This can be heavily simplified by just setting the flag to ignore files
that are of unknown type. That leaves the ignore list just with files we
do not want to be formatted on purpose. Prettier is already quite smart
in which files it acts one (like following `.gitignore`, output files
defined in `package.json` etc.). Thereby, actually nothing remains to be
ignored.

The `CONTRIBUTING.md` file was now actually formatted by this, as it is
included in the overall list of files to format. Which is correct.