Chore: simplify script pattern for linting
It is arguably better to keep the scripts in the `package.json` quite
simple and concise. It can also be argued that is better to call
linting, formatting, ... tools with all files and configure them about
which files to actually care about or which to ignore (if needed). That
keeps the configuration of the tools in a central location and the
scripts basically just call these tools.

I'm currently unsure if/how the `*.config.{js,ts}` files should not be
somehow included by linting. Would be beneficial, but currently they are
not part of the `tsconfig.json` scope.

Additionally this ensures that also ESLint warnings count as "errors".
For CI/CD, VCS-hooks it is not really meaningful to have lints of
different severity levels. Either they need to be handled or not.
Else, reports that must not be handled can pile up. Leading to
desensitized developers which will simply ignore the output. New reports
will be missed. In result the value of these reports becomes void. So
either they count or they don't.