Chore: streamline git hooks with package scripts
This makes the scripts in `package.json` slightly more powerful by let
them accept additional arguments conditionally. This way, they act as
before then called without arguments (i.e. working on the whole
repository/all files). But can be called with a specific (list of)
file(s). This enables for example to directly use these scripts within
the git hooks and provide a list of staged files. Before that, it was
necessary to keep the commands in the hooks somehow in sync. This
spreads the logic and is simply unnecessary.