Chore: remove Lefthooks comments in configuration
Comments are a tricky topic in clean code. Anyhow, it is not tricky to
say that comments that explain configuration options of a tool are no
good comments. Therefore the documentation of the tool/library/... is
for. In best case the configuration format is powerful enough/well
supported so that an IDE can help with getting to the documentation
(e.g. typed scripts, JSON schemas definitions, ...).