Feat: add new user feedback section component
The component is purely visual for now. The current plan is to use the
attached button identifiers to trigger events with the tag manger of the
user tracking setup. This might change by future iterations.

Furthermore the component feels interactive right now. Like nothing
happens when a button is clicked.