Fix: type gap for summary page
There was an issue with the data stream for the planning overview. The
recently added months created a little gap. That is that down the
stream, months have been reduced to only consist of the variant/type and
the maternity protection. But suddenly at the very bottom it was
expected to carry all the money values again. It just worked out,
because all these values were declared as optional. While during
creation of the data is never undefined.
This also worked out, because in the selectors and combiners that
transform the data from the store, there were no strong function
signatures in use. This might easily happen because the nested levels of
the type definition did not match the level of functions operating on
it. Adding these signatures makes the type-system complain again. This
is one of the known weaknesses of TypeScript in general.
Furthermore, there was an easy type mix with the `Month` type form the
store (well not directly). Which made things a little more shady.

So there was a need to name this type which represents a "month" (to not
use `Pick<...>`) everywhere. In lack of a clear domain/feature language
atm. it has been called `Lebensmonat`. Mostly to distinguish it more
clearly from the `Month` of the underlying data structure, which
represents a slightly different concept.
I'm not really happy with this naming. But I also don't think it is bad
and it should currently help the reader to easily grasp and distinguish.

Furthermore, there is this type coercion going on down the component
stream. Components on lower levels of the tree reduce the definition of
the `PlanungsdatenFuerElternteil` for the scope of the details sections.
Which makes sense and works. Though, might be a hint that the overall
feature is no more well defined and might have to diverge. Or the type
system has to represent this better.
As the execution environment is still just based on JavaScript, awkward
behaviors are possible due to this coercion (works properly in other
languages).

In the end this lead to a changed data structure at the leaf component.
In exact that the money values can no more be undefined. Therefore, the
component had to change a little. The presence of a money value was now
determined by if it is not a zero number.

As a final note:
The types in the `planungsuebersicht` folder have the scope of this very
feature. They represent the data based on the view of this feature. To
make all the little parts of this feature (components, hooks, selectors,
function) work nicely together on this data and to support the reader,
these types are centrally defined. This way we have very explicit data
relationships in contrast to weak implicit type correlations.
That approach only works as long as the feature is scoped well enough.
Features getting too big will suffer from types which serve any
purposes. The feature of topic is probably right at the edge here.
Because it provides to different views of the same underlying data.
Though, this all might simplify and resolves itself with a new domain
model in general.