Feat: align variant type for summary page
When the feature of the summary page was started, the `None` variant (of
the underlying `ElterngeldType` of the store) was excluded. So it was
the idea that form the perspective of this feature, the possible
variants are just the three others. This was resolved in the data
transformation functions.
Adding the details view to the feature has actually changed that
perspective. `None` is still a possible variant. But for the details
view, this variant will be skipped or just not shown.

Getting this clarification, it was possible to unify and align the usage
of the variant for the whole feature. Thereby, there is no more awkward
split between the different "sections" of this feature. It also clean
ups the partial leak of types from the store across the features type
layer.

There is hope that all this simplifies a little, once the central type
system of the data got a strong domain to build upon. Which should not
mean that features must/should not represent their own views. But these
layers should get thinner.