Chore: minor improvements for Plangungs* components
Just little boy-scout stuff.
The property `sumMonths` was not easy to understand what it should do.
I hope with the new name it is a little more expressive across component
boundaries.

Resolving some "unsafe" and hard to read conditional array access.
Treating the array of data for each parent as an array no matter the
length makes the code much simpler. With the current data structure of
a pure array, everything is just assumptions. Not even the first parent
must exist.
I guess this can be resolved even better with a future strong
representation in the type system. But for now, this is more streamlined
with all the other components treating the variable number of data for
each parent as a pure array without any implications of the exact
length. Makes the code much more simple and robust.

Furthermore there is a tiny change involved in regards of more extensive
index variable naming. The simple and often used variable names of
`index` or `i` to represent the index in a loop is quite common. But it
also expresses not much. This works fine in many scenarios where the
loop is very simple or just quite mathematical without further context.
But at least when loops get nested and we mix `index` and `i` with each
other to resolve name conflicts, this gets dirty. Using more descriptive
names leads to more line breaks, but arguably makes the code overall
better to read.