Feat: try to catch spirit of summary details
There has been a discussion about the current design concept of the
details per month on the summary page. We came to the conclusion that it
is not 100% refined. Probably even a little confusing here and there.
This confusion was also visible in the code of the component, which was
hard to grasp and contained some edge cases that remained unclear.
In absence of the original designer we agreed on minor improvements. And
that probably some iterations will follow.

But most importantly we better aligned on the underlying structure of
the table cells and its exceptions:
There is a visual representation of the variant at the start. Though,
there is still an exception on how maternity protection months or months
with the `None` variant appear in contrast to the rest. Furthermore
there is a list/stack of the following three money values:
  - elterngeld (skip if none, bold if only money this month)
  - netto einkommen (skip if none, bold if only money this month)
  - verfuegbares einkommen (skip if only elterngeld or netto, always
    bold)
The implementation of the component tries to fully represent exactly
this written/verbal description. So the code falls in line
understanding/concept of the design.

In addition to that, we discussed the cases of "incomplete" months. This
includes months that have no variant selected (also not even `None` ->
future topic) and optionally include some net income.
This also presented the new implementation with the need to fill missing
months for the parent that takes no months when the other still does
(awkward description). This brings the responsibility to the table
component to fill these cells, so they are not empty. This worked before
by "accident".

Another minor change has made it to the commit which is the flattening
of the month component properties. Having a single property that holds
an object is one unnecessary level of nesting, leading to more "complex"
code that is "hard" to read. A component that is called `*Month` that
takes a `month` property kinda makes sense. But in turn it could also
just directly take all properties that make up this month directly. This
is actually a case by case decision. But if the single prop gets
directly spread in the function body this is a good indicator, as in
this case.