Chore: minor improvements for PlanungsdetailsTable
The order of code in a file matters a lot to the reader. If navigating
to a file not by language reference (should mean no jump-to-definition or
similar) but by selecting this file, the reader has some expectations.
In best case was the folder/file/module/... structure already quite
communicative. Then it is not really helpful for the reader that at the
very top he is faced with utility functions first. At the very top
should be placed what this specific file/core is about. And below follow
the "children" of what ever it is.
This succumb the restrictions of the language and compiler.
Traditionally something had to be defined first in code before it could
be used. This strong relationship has been resolved by many compilers
these days on a textual code level. So it is possible to write things
much more decoupled from these relationships and order them based on
reader purpose (notice that in TypeScript imports still have to come
first, so there are some restrictions for constants).

Furthermore there are functions in every programming language. And these
are easy to identify and process, especially in TypeScript as these are
declared by the keyword `function`.
There is also the concept of lambda/closure/anonymous functions in many
languages. These serve specific purposes and have special behavior
(though vary between languages). They get quite abused these days in
Java-/TypeScript these days where the function context does not matter.
But they lack one strength of regular functions: expressive code.
A function with a name is better to read as a variable declaration with
an assignment to a callable closure.

Similar arguments apply to the definition of arguments of
a (what-ever-kind-of) function. A pure list of named arguments is still
the most simple and straight forward way to declare arguments. Using
objects is a good mechanism to solve certain scenarios like a big group
of optional arguments. But also comes with additional complexity, also
for the caller.
The strong naming of variables on the caller side is sometimes an
argument. But can also be turned into a contra argument easily as
visible in this exact code where property name mapping is necessary.
Usually readability of pure function arguments is very good on the
caller side due to modern IDEs and inlay hints, where naming derives.