Chore: extract PlanungsdetailsTable props
This is currently the used style over the whole codebase. So this is in
first place about consistency. But there are also good arguments why
this style provides some cleaner code. In terms of lines-of-code, this
is actually in most cases 100% the same.

We might wanna consider to agree on a specific style and enforce that by
a linting rule.