Chore: remove obsolete documentation
In regards of working towards becoming open-source, it is important to
review the existing documentation (which is not much). It has not been
touched at all since we took over the codebase. In result it is quite
obsolete and does no more match in many points. And wrong information is
worse than no information.

To address this with a first step, the CONTRIBUTION file got removed. We
do not work with such a concept at the moment and we need to review it
first when it becomes relevant again. Furthermore, the README is
out-of-date. We switched the build tool, test runner, move away from
SCSS and many other technical parts. We also initially merged multiple
repositories of ]init[. Currently we don't have proper documentation
here. So a minimal README just includes the most basic list of commands
to run for this project from installation to production build.