Chore: tidy up package.json
We currently do not really maintain a version. Especially no semantic
version. We do releases in our Pipeline to automate the production build
in a clean environment. The release names are still unrestricted and
currently use a date string.

The `eject` script is a leftover from the create-react-app time and is
obsolete by now. It would actually fail to execute.

The browserlist configuration got removed because it is no more used.
Our new build tool `Vite` does not read these values. So wrong
information is worse then none. I opened a ticket for us to get a clear
perspective of browser support.