Feat: add "effect" to user feedback component
To give the user some feedback that clicking on any of the feedback
buttons did anything successful, the component will switch its state
to change the view as being completed.

This includes changes to the stylistic deviation of the semantic heading
element as agreed with design.

Furthermore this makes use of the improve `Button` component,
removing the need to overwrite styles.