Revert "Fix(a11y): let screenreaders ignore +/- buttons in counter field"
This reverts commit 1a6f6e8d714bac9c8e91893a1496e9edb5d7698d.

Hiding the buttons causes unexpected trouble with our tests. Tried a
few things to no avail. Maybe we can work around it, but it won't be
pretty. There must be a reason why testing becomes harder...

I'm no longer sure, we should exclude the buttons for a11y tech. Maybe
we take this as a hint to remove the buttons for everyone.