Feat: adapt labels, hints, etc. for Mini-Job limit
Important notice:
We agreed with the project partners to only update the textual part of
this change. The limits for the actual calculation remain untouched.
That means the calculation will be off for between 520 till 538 Euro.
This is because the overall calculation causes many issues with the
continuously changing limits. The current approach is not designed for
this. The future of the calculation has to be rethought in general.