Feat: kick-off reworked planer feature with minimal domain
It is the goal of the project to rework the planer feature of the
application with a new approach. This is taken as opportunity to
clean up some legacy code and lay down a new foundation.

With the goal of a more "screaming-architecture" and some clean code,
the approach of feature directories was introduced. This splits the
application code into features. This helps developers to manage the
increasing complexity of a growing application.
The "old" part of the application remains as is for now. It will also
transition into the components and functions that are re-usable cross
features.

Furthermore, a Domain-Driven-Design inspired approach will be used (at
least for this feature). The goal is to create a solid base of data
structures, operations and specifications representing the domain in
a neutral manner. It uses the ubiquitous language of Elterngeld.
While intending be designed in an agnostic manner, the domain will be
finally consumed/provided by a user interface. With the current
architecture of a client-only single-page-application, both live close
together. As many web component frameworks (including our chosen
ReactJS) use a functional inspired approach, the domain is designed in
functional manner too. That means read-only data structures that are
decoupled from their related functions used to create, manipulate or
inspect them.

The here added domain is a very first collection of data structure
descriptions, which are needed to construct a plan for Elterngeld. A new
key concept is the generic part of the Elternteile. It is an approach to
put the fact that there is a varying number of Elternteile into the
type-system, enforcing all related code to be conform. This is an
improvement for the weakness of the old code.

Furthermore the domain is German. Therefore a lot of the related code is
German too. This is always complicated and never perfect. But it is not
feasible and not in the spirit of the ubiquitous language to translate
everything to English.