Feat(planer): add operations to create an initial plan
The first obvious use-case of the new Planer domain is to create an
initial plan.

This introduces the new anticipated directory structure within the
domain that puts `operation` functions next to their related data
structure.
Furthermore, operations act in first place only their level of data
structure. That means there is not a single top level function which is
responsible to create a whole plan with all included behaviors.
It would probably be split into multiple functions anyway to control the
complexity level, produce clean code an maintain the testability. The
Slicing based on data structures helps to keep them somewhat decoupled.
While they are of course strongly related, it keeps the dependencies on
the details level low. A Lebensmonat must not necessarily now all the
details of its child Monate.

In addition to make the use-case meaningful, there are some initial
information necessary. These are represented by the `Ausgangslage`. This
was necessary, because the data structure of `Lebensmonate` is meant to
only include planned `Lebensmonat`e. So an initial plan would be
actually empty. But taking into account that there are Monate with
Mutterschutz, this needs to be addressed during the initial creation.

This new "root configuration" of the plan is also used to resolve the
generic pattern of the `Elternteil`e a plan consists of. It dictates how
many `Elternteil`e there are. It strongly links this to the
`Mutterschutzempfaenger`, which is already a first example why this new
concept is very powerful to enforce inherently correct plans.

Notice that the type-system of TypeScript is very strong. But the
compiler is unable to derive all types. Like for example for return
types conditionally based on return types. There are open issues for
certain weaknesses of the compiler. Till then, specific type
enforcements are necessary. Special caution is required when writing
such code.