Feat(planer): add domain operations to safely list data
This is in first place to help with some type system issues while
reading data structures of the domain.
While the data types of the domain can be just plainly read, there are
issues in regards of the `Record` type by TypeScript. Further details of
the actual problem are annotated in the code itself.
So these new operations are not really mandatory in first place, but
help to make the interactions with the data more flawless and type-safe.

There has been a `common` directory added to the domain. While the root
directory of the domain is kept simple and flat on purpose (e.g. no
nesting of object-values, entities, ...), there will be a bunch of
complementary utilities used within the domain. So they are not part of
the domain from a non-technical point of view. This will also be the
place were patterns like specifications will be defined.

The new methods to list the entries of a record are hosts of unsafe
code. They are written in careful manner to handle everything correctly,
so that the caller side can remain safe and hassle free.