Feat(planer): add domain operation to choose option
This new operation/use-case allows to actually take a decision and
choose an Auswahloption for an Elternteil within a Lebensmonat. As
already described earlier, the operations is split across the affected
domain structures. Each level handling its own logic.

Due to the design to allow for an initially empty list of plannend
Lebensmonate, or in fact a partial set of only planned Lebensmonate,
there is the need to create new Lebensmonate when chosing an
Auswahloption. To do so, this requires the information of the
Ausgangslage. Therefore a new data structure called Plan has been put on
top of Lebensmonate. This new root aggregate entity is responsible to
hold all relevant information needed to have for consistent operations
with invariants. It is known from the initial design phase of the domain
model that there will be more need for this.