Test(planer): separate imports for in-source tests
For in-source testing, there are frequently additional imports needed.
Should mean, the tests need something that is not needed by the code
they are testing.
Using straight forward imports will "pollute" the import section, and
thereby the dependencies, of a module also for production. Tests are no
more an add-on. This isn't critical, but can definitely be better.

The chosen solution makes use of dynamic imports. But dynamic imports
are asynchronous. It is not possible to define self-calling functions in
a conditional block. Therefore `describe` blocks are used for this.
While they might usually not be necessary for in-source testing, they
help to add an asynchronous wrapper around the test cases. Within that
wrapper all test only dependencies can be imported. That is also true
for type-only imports of the source code.

The big downside of this approach is that it can't be easily enforced.
IDEs will usually promote/automatically "regular" import(s) at the top
of the file. There might be the opportunity to write a custom linting
rule in future to help enforcing this concept. Till then, it is the
responsibilities of all engineers to follow this (or discard it
altogether).