Feat(planer): integrate Elterngeldbzüge into choosing an Option
There is a separation of the calculation of the Elterngeldbezüge and the
planning. The first are calculated taken the whole plan as input and put
out an entry for every Lebensmonat, Elternteil and Variante. The
approach is to keep these data structures separated and integrate them
back together based on operations. The first part of this is the storing
the known Elterngeldbezug when choosing an Option. This integrates the
Elterngeldbezüge into the plan for specific data points. Further
integrations will follow.

There arises the need to provide test data generators to efficiently
create test data with readable test cases.