Feat(planer): add operation to update Elterngeldbeträge
As described for the operation to chose an Option, the calculation of
the Elterngeldbeträge is a separated domain. There are different
triggers when they are updated. Most importantly, the Bezüge of already
chosen Optionen will change based on new calculations. Therefore, it is
necessary to update these. The Plan entity is again responsible for this
invariant.

This includes some "unrelated" improvements to various tests. The intend
is to use some first versions of test data generators. In a minimal
local/duplicated approach. It helps already a little bit to make the
test cases more readable.
The approach right now build on top of removing object keys by named
function signatures. This allows for terse data definitions within
a single line of code. If the developer has an editor that supports
inlay hints of function signature parameters, this is as expressive
writing out the keys in actual code.