Feat(planer): introduce specification pattern to domain
Adds a first version of a custom implementation for the specification
design pattern. It is intended to be used for in-memory validation
and data retrieval in a composable variant.
The included concept of violation messages is intended for user
interactions like disabled hints or error messages. Internationalization
and further details are not yet implemented and left open for future
additions when necessary.

The first added specifications encapsulate the business rules for Monate
mit Mutterschutz. The levels of specifications are meant to work similar
to the operations: per data structure. Though, this is not always
feasible and meaningful, making business logic harder to grasp. This is
for example the case for these first specifications too where the
Lebensmonat and Lebensmonate level are skipped.
It is another example of where the Plan is used for invariants. Further
integrations of actually integrating the specifications will follow.