Feat(planer): evaluate specification when choosing Option
This is the first application of a specification for in-memory
validation purposes. It is used to validate if a Plan is still gültig
if a certain Option is chosen. If not, the resulting Plan will be
discarded and the messages of the specification violations are forwarded
in an error state.

This introduced another domain common implementation: the `Result`.
Please see the inline documentation for further details of this concept.