Feat(planer): add domain operation to determine Auswahlmöglichkeiten
This closes the circle of operations to create a Plan by providing
information about possible Auswahlmöglichkeiten for a specific
Lebensmonat and Elternteil. It uses the Plan root entity for the
invariants again.

It is the goal of the planning feature to guide the user as much as
possible and prevent him from creating wrong Pläne. To train the
understanding, rules should be explained for each situation.
Thereby, it is the strategy to provide the user with a list of possible
Auswahlmöglichkeiten for each point he has to take a decision. If
an Auswahlmöglichkeit is "available" but its selection would lead to an
invalid Plan, it gets annotated as disabled with some information why.
To achieve this, each Option is actually selected "virtually" and the
returned result is interpreted.