Feat(planer): add domain specification for parallel Basis
This new business rule leads to an extension of the specification
implementation. It adds composability where multiple specifications for
the same generic instance type are evaluated altogether. Notice that
with the current design, there are no further logical operations (like
not, or, xor, ...) possible due to the violation messages.

There is a technical debt evolving which has to be resolved in future.
Parts of the domain/application are agnostic in terms of how many
Elternteile there are. They are written with algorithms which work for
any number. And because there are usually no more invariants to take
care of they should not be generic of the Elternteile or Ausgangslage.
That makes them more simple and also allows them to be constants.
But currently that does not work well. I haven't figured out a good way
to represent this type that just works in all scenarios because how type
unions and type coercions work in TypeScript. In result, some
functions/operations can't be non-generic with a less strict type. And
also specifications must always be invoked functions instead of
constants to push up the generic resolution. This feels awkward because
it enforces to use generics where they don't seem necessary (and should
not).