Feat(planer): add concept of verplantes Kontingent to domain
This adds the data structures and operations to count the chosen Options
within the Lebensmonate of a Plan. This will be complemented by the
verfügbares Kontingent. Besides UI purposes, this will also result in
specifications building on top to enforce the business rules.

Notice that not all operations go down the tree of data structures. Also
not all operations go up. It is a case by case decision to break up
complexity.