Feat(planer): add specification to check Kontingent
Notice that this is the first specification actually implemented as
a class. That is due to its dynamic violation messages. Also would
a break down into smaller specifications also for lower data structures
actually not help, but rather make it more complicated.

The function on top of the class replaces the need to call the
constructor itself. That's for more uniform code. In best case, most
specifications are just constants. Especially if they are not
configurable. Those which are, should generate specifications by
function parameters. The classes themselves are only for the base
implementation that provides the composability plus base methods for
interactions. Else only specifications with special needs should need to
implement a class.
But because of the current technical debt of generic pass-through,
a constant version is not possible here.