Feat(planer): add specifications for Partnerschaftsbonus
Kick-off the complicated rules and logic around Partnerschaftsbonus.
This adds the validation part for the first and most simple aspect of
it.

This also adds a fabric method that allows to create specifications for
properties of a type. This helps to use/compose specifications for high
level data structures, but keep the actual specifications implemented
for the lower level data structures they are actually operating on.