Feat(planer): add automatic Partnerschaftbonus selection
To support the user while creating a Plan, but also to seal the concept
of an always valid Plan, the operation to choose an Option has become
more advanced. If one Elternteil chooses the Partnerschaftsbonus Option,
also all other Elternteile get this Option chosen. And the other way
around, if any Elternteil changes its choice from Partnerschaftbonus,
all other Elternteile are reset.
Thereby, the latest specification of only parallel Partnerschaftsbonus
is always automatically guaranteed.

The implementation is not beautiful. It tries to use composition of
functions. The main function determines which case it is and delegates.
The functions specific per case then decide which Elternteil needs to
chose what. This breaks the complexity level by level. Though, the end
result does not seem optimal.

This adds a set of operations to reset data structures. This comes
useful as it names the concept explicitly and wraps the tiny logic. So
it can be used directly, without hassle and good readability. Also there
will be more use-cases in future.