Feat(planer): extends automatic Partnerschaftsbonus selection
This continues the user support to satisfy the business rules of no or
at least two Lebensmonate with Partnerschaftsbonus. Implementation wise
is follows the same approach on the Lebensmonate level, as already done
before for the Lebensmonat data structure.

This had some side effects on tests which only provided some
mocked/minimal Elterngeldbezüge. This is kind of a technical debt,
because there is still no proper test data fabric setup.