Feat(planer): extend automatic Partnerschaftsbonus selection
This fixes the automation algorithm to also unchose Partnerschaftsbonus
where needed. The test cases are not complete, but test the most
prominent and descriptive cases. The implementation is a very generic
algorithm that works very solid for all various possible cases.

The operation on the Lebensmonate level starts to become really complex
and longish. Though, I haven't found a good way to break it down into
multiple modules. Especially not with the limited access modifiers
TypeScript provides. If this operation continues to grow, we have to
find a way.