Feat(planer): kick off user interface with store selector
This commit finally starts the work on the user interface for the planer
feature. But as a first building block, it needs a service which
provides the state management and the methods to manipulate the data.
Kind of similar to DDD services.
Anyhow, the actual "service" comes soon. The committed work only
provides a Redux store selector to create the Ausgangslage data
structure. It is the starting point (well, thereby the naming...) for
a Plan which comes later.

This also starts the export-import game between the domain module(s) and
the user interface module(s). It is the idea to keep them as independent
as possible and shape explicit relations. Therefore, the domain exports
explicitly only what is needed from it. And the user interface service
imports at a single location what it needs from the domain. This creates
a straight relationship.