Feat(planer): add selector for Elterngeldbezüge
As the generic game is hard to play in this part of the application,
this selector always gives back the Elterngeldbezüge for two
Elternteile. This works always as the calculation does so too. In the
later part of the application which reads the data it will cause no
harm perse.