Feat(planer): establish service hook for UI
This service is the hub for the whole user interface of the planer
feature. It provides the data management and exposes the bound domain
operations to act on the data. Furthermore it includes a few logical
aspects to keep certain data in sync or what to do on an error.

This includes a bunch of definitions for callback functions. These are
already bloated up a little with intend to use them soon within the
components that act on each level of the data structure.

This fixes a file naming which incorrectly used hyphens.