Feat(planer): add Pseudonyme for Elternteile
Theoretically this could be seen as a UI only feature. Though, it is
still part of the ubiquitous domain language. Though it is only loosely
coupled to the rest of the data structures and is not part of any
operation. Anyhow, the pseudonym is often used in the UI as replacement
for the generic Elternteile type. The latter is necessary from
a technical point of view and can't be simple replaced by using the
Pseudonyme directly.

Because the Pseudonyme were added as required property of the
Ausgangslage, this caused a lot of changes everywhere in the tests where
Ausgangslage data structures are created. This is another hint that good
test data factories are needed. Though, the advanced generic types and
type discriminators show that these factories are not trivial.
Especially if remaining terse for readable test cases.