Feat(planer): replace old with new planer
Actually start to use the new planer feature. Some minor features are
still missing and will be added.

This shifts some features. Things like blocking the planer, page
navigation etc. are not part of the planer feature itself. It is part of
the responsibility of the component were it gets integrated.

A lot of code will become delete soon by this.