Chore: remove Prettier plugin from ESLint
This does NOT remove the compatibility integration between both.

There is a fine line between code formatting and linting. Many rules in
ESLint check for some kind of format and is capable to automatically fix
them (read: format). This is for example the case for sorting of
imports.
Anyhow, there is some base formatting done with Prettier. Prettier
formats each file. It is kind of an abusement to add all Prettier
formatting rules as linting rules with automatic fixing.
This becomes especially annoying during interactive development. Because
many editors can integrate with ESLint and show diagnostics for the
errors. But for formatting this creates tons of noise all the time while
typing. Formatting usually just applies automatically when saving
a file. Or finally with the pre-commit hook.