chore: add linting for commit messages
This commit message should not become a pleader why commit messages are
important and in return why linting them makes sense.

The configuration takes the conventional commit rule set as base as it
is one of the most widely used commit message format in the open source
community. It is a very sane "default" without much arguing about
configuring each rule individually. Anyway, this has been "merged" with
the default configuration of our company templates. But only a single
rule that has no conflict with the preset was taken over.
It should be mentioned that conventional commits aim for an automated
changelog. This is currently no goal for us and does not fit the trunk
development approach. But anyhow, this is still a very sane standard
even without this "feature".