style(monatsplaner): improve responsiveness of gewählte Option
The horizontal width requirements of the Planer are a little tricky as
no wrapping is possible. Also the wrapping of cells in the grid is often
not possible. Therefore some conditional elements are necessary to
reduce the displayed content.
In this case it is a little tricky as CSS styling features do not
provide sufficient features to implement this fully intrinsic. See the
comments for further information.