feat(monatsplaner): show Zeitraum for each Lebensmonat
This commit might look bigger than it should be. That is for one part
due to prop-drilling (which is arguably still fine). But for the
majority this is due to the lack of test data generators and the new
required data field. This is still a technical dept to resolve (though,
still tricky).