feat(monatsplaner): integrate tracking for Partnerschaftlickeit
The tracking itself still remains separated from the feature of the
planer itself on purpose for now.
That includes the calculation of the
quotient itself which is now based on the domain model of the planer.
Though, the calculation is not yet very "generic" feels a little clumsy
by acting on the two parents directly.
But the triggering of the tracking is done by the planer again, as it is
based on events/actions there. This might change when shaping the
feature interactions in future.

This commit starts now to finally remove old code that got
adapted/replaced with new code now. Like the integration of the tracking
here got removed from the old Monatsplaner as it is now available in the
new Planer.
There is still a long backlog of old/dead code to remove. So despite
continuously removing replaced code commit by commit, a bigger cleanup
will be necessary soon.